Release 0.1.1

+++ src/system.rs
@@ -1,5 +1,5 @@
 use std::fmt::Debug;
 
 pub trait System: Debug {
-  fn step(&mut self, dt: f64);
+  fn step(&mut self, dt: f32);
 }
